streamlit_app.py
"""
Streamlit App for Snowflake
A Streamlit application template configured for Snowflake deployment
"""

import streamlit as st
import snowflake.connector

st.set_page_config(
    page_title="My Streamlit App",
    page_icon="🚀",
    layout="wide"
)

st.title("🚀 My Streamlit App")
st.write("Welcome to your Streamlit application on Snowflake!")

# Example: Connect to Snowflake (configure as needed)
# @st.cache_resource
# def init_connection():
#     return snowflake.connector.connect(
#         user=st.secrets["snowflake"]["user"],
#         password=st.secrets["snowflake"]["password"],
#         account=st.secrets["snowflake"]["account"],
#         warehouse=st.secrets["snowflake"]["warehouse"],
#         database=st.secrets["snowflake"]["database"],
#         schema=st.secrets["snowflake"]["schema"]
#     )

# Add your Streamlit app code here
st.write("This is a basic Streamlit app template for Snowflake deployment.")
