requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
snowflake-snowpark-python>=1.0.0
snowflake-connector-python>=3.0.0
